ai-assistant-service/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
opensearch-py==2.2.0
openai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
numpy